Fully support KMP for stream-result and stream-result-call
