+++ build.gradle.kts
@@ -5,6 +5,7 @@ plugins {
   alias(libs.plugins.kotlin.android) apply false
   alias(libs.plugins.kotlin.serialization) apply false
   alias(libs.plugins.kotlin.multiplatform) apply false
+  alias(libs.plugins.kotlin.atomicfu) apply false
   alias(libs.plugins.ksp) apply false
   alias(libs.plugins.nexus.plugin) apply false
   alias(libs.plugins.spotless)
