+++ gradle/libs.versions.toml
@@ -1,4 +1,5 @@
 [versions]
+atomicfuVersion = "0.25.0"
 streamLog = "1.3.1"
 androidGradlePlugin = "8.6.1"
 androidxActivity = "1.4.0"
@@ -11,6 +12,7 @@ androidxNavigation = "2.5.0"
 androidxTest = "1.6.1"
 androidxJunit = "1.2.1"
 kotlin = "2.0.20"
+atomicfu = "0.25.0"
 dokka = "1.9.0"
 kluent = "1.73"
 kotlinxCoroutines = "1.9.0"
@@ -30,6 +32,7 @@ mockito = "5.13.0"
 mockitoKotlin = "4.1.0"
 
 [libraries]
+atomicfu = { module = "org.jetbrains.kotlinx:atomicfu", version.ref = "atomicfuVersion" }
 stream-log = { group = "io.getstream", name = "stream-log", version.ref = "streamLog" }
 androidx-appcompat = { group = "androidx.appcompat", name = "appcompat", version.ref = "androidxAppCompat" }
 androidx-annotation = { group = "androidx.annotation", name = "annotation", version.ref = "androidxAnnotation" }
@@ -72,3 +75,4 @@ spotless = { id = "com.diffplug.spotless", version.ref = "spotless" }
 dokka = { id = "org.jetbrains.dokka", version.ref = "dokka" }
 nexus-plugin = { id = "com.vanniktech.maven.publish", version.ref = "nexusPlugin" }
 kotlinBinaryCompatibilityValidator = { id = "org.jetbrains.kotlinx.binary-compatibility-validator", version.ref = "compatibilityValidator" }
+kotlin-atomicfu = { id = "org.jetbrains.kotlinx.atomicfu", version.ref = "atomicfu" }
