+++ stream-result-call/src/commonMain/kotlin/io/getstream/result/call/ReturnOnErrorCall.kt
