+++ stream-result-call/src/commonMain/kotlin/io/getstream/result/call/ZipCall.kt
@@ -18,19 +18,19 @@ package io.getstream.result.call
 import io.getstream.result.Error
 import io.getstream.result.Result
 import io.getstream.result.call.dispatcher.CallDispatcherProvider
+import kotlinx.atomicfu.atomic
 import kotlinx.coroutines.async
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.withContext
-import java.util.concurrent.atomic.AtomicBoolean
 
 internal class ZipCall<A : Any, B : Any>(
   private val callA: Call<A>,
   private val callB: Call<B>
 ) : Call<Pair<A, B>> {
-  private val canceled = AtomicBoolean(false)
+  private val canceled = atomic(false)
 
   override fun cancel() {
-    canceled.set(true)
+    canceled.value = true
     callA.cancel()
     callB.cancel()
   }
@@ -40,16 +40,18 @@ internal class ZipCall<A : Any, B : Any>(
   override fun enqueue(callback: Call.Callback<Pair<A, B>>) {
     callA.enqueue { resultA ->
       when {
-        canceled.get() -> { // no-op
+        canceled.value -> { // no-op
         }
+
         resultA is Result.Success -> callB.enqueue { resultB ->
           when {
-            canceled.get() -> null
+            canceled.value -> null
             resultB is Result.Success -> resultA.combine(resultB)
             resultB is Result.Failure -> getErrorB(resultB)
             else -> null
           }?.let(callback::onResult)
         }
+
         resultA is Result.Failure -> callback.onResult(getErrorA<A, B>(resultA).also { callB.cancel() })
       }
     }
@@ -76,14 +78,14 @@ internal class ZipCall<A : Any, B : Any>(
     val deferredB = async { callB.await() }
 
     val resultA = deferredA.await()
-    if (canceled.get()) return@withContext Call.callCanceledError()
+    if (canceled.value) return@withContext Call.callCanceledError()
     if (resultA is Result.Failure) {
       deferredB.cancel()
       return@withContext getErrorA(resultA)
     }
 
     val resultB = deferredB.await()
-    if (canceled.get()) return@withContext Call.callCanceledError()
+    if (canceled.value) return@withContext Call.callCanceledError()
     if (resultB is Result.Failure) {
       return@withContext getErrorB(resultB)
     }
