+++ stream-result/src/commonMain/AndroidManifest.xml
